Updates & Simplifies midi2piano.py
Before anything, I am very much not a coder of any form and barely know how Github works, so apologies if there are any errors here or in the way I am doing things.
Now, this changes two things about the horribly outdated tool that is midi2piano.

First, it allows for a user to simply put midis into a "midis" folder which it will process and convert into .txt files, rather than the user having to manually paste their text. This was entirely done by Alainx277#0816 on Discord a few months back, so full credit to them.

Secondly, it updates the settings to actually match how the current instrument system works on Bee, which supports 300 characters per line and 1000 lines, but by default the tool would only do 50 characters and 300 lines. 

And that's about it! If there are any issues or dumb things I've done, do let me know and lend a hand, if you could.